chore: use native token payments on e2e tests (#245)
This adds coverage on tests to use native token (ETH) as payment
methods.

In order to do so, it refactors the `run_e2e` internals so it's run from
a list of `TestConfig`.

With this, `run_e2e` will run the transaction list with both PREP and
upgraded accounts, in both ERC20 and native payment methods, while
enabling future test configurations.