refactor: rm `sendAction` (#277)
removes `sendAction` from the public rpc and makes it a private function
on `Relay`. additionally, changes the signature slightly, since we have
all the info we need in `SignedQuote`.

a note here would be that `SignedQuote` kind of is a weird name now...
same with `Quote`. we should consider renaming it (as it is really a
quote + a user op and more)

Closes #216