feat: optional nonce (#283)
i've moved `PartialUserOp` out of `sol!` so we can make some fields
optional. this is still technically just temporary until we address
#215, but this allows us to make `nonce` optional.

if the nonce is not specified, we just fetch the next nonce for the
sequence key `0`. this should work for now, but if we want replayability
on the userops, we should use a multichain sequence key.

closes #207