feat: tests for `TransactionService` (#293)
Adds utils for testing the `TransactionService` separately and a simple
test asserting that events and metrics are getting handled correctly.

The tests are written by using `MockAccount` which is just a random PREP
EOA which is capable of producing `RelayTransaction`'s which can be
submitted directly into `TransactionService`.

For now the test simply asserts that we can handle some concurrent
transactions