feat: add accounts to `AccountRegistry` outside of prep initialization (#295)
on top of https://github.com/ithacaxyz/relay/pull/292

Accounts were only being registered during the PREP intialization. This
PR ensures that every authorize of an admin key OR webauthn must come
with a ID signature, so it can be added to `AccountRegistry`.

`AuthorizeKey` takes an extra optional argument `signature`
* `Some(KeyHash)` -> if it's an admin OR webauthn key AND we know the
EOA address (prep or upgraded)
* `None` -> if we're calling `prepareAccount` (we don't know the prep
address) OR it's a session key which is NOT webauthn


Follow-up will integrate `AccountRegistry::appendAccount`