feat: estimate fee by setting `userOp.paymentAmount` to 1 (#356)
Most tests were able to pass with 10k user op gas buffer. 

The `session_key_pre_op` required 40k so that's all we can decrease to.
I wouldn't change the staging relay just yet.

The caveat is that the user would need to have ERC20 if he wants to pay
with it. It's not the case with ETH, because we can easily state
override it (PR includes it). But overriding ERC20 is trickier.

The preops should be constructed with ETH as `paymentToken`