feat: rework tx service (#358)
closes #328
closes #336 
closes #337 


changes the architecure of the the tx service <-> signers

this makes signers identifiable, introduces a shared channel through
which the signers communicate back to the service, this makes it easier
to process events sent to the service from the signers. and this way we
can map events from signers to the tracked signers.

---------

Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>