fix: limit number of pending transactions for tx service (#370)
Right now we might broadcast unlimited number of transactions which is
not great because we can't expect all of them to be handled correctly.

This PR limits number of transactions that may be pending.

This is done by tracking `last_confirmed_nonce` and by introducing a new
tx state — `UnsentTransaction` which is a transaction with assigned
nonce waiting to be sent.