feat: pause signers if balance is too low (#389)
Adds logic for pausing/unpausing signers if balance reaches certain
values. Current implementation will pause the signer if it can't afford
spending 30M gas per the current gas price estimate. This should be a
reasonable bound enough to process any pending transactions that may be
still in progress.

Balance metrics are still emitted separately by `BalanceCollector`. I'm
not sure if we should move this logic into `Signer` and remove the
`BalanceCollector` as I assume we might also want to emit metrics for
token balances which don't really belong to signer as it doesn't know
about tokens that we support.

So imo if we want to collect and record both native and tokens balances
those should happen in a separate `BalanceCollector`, otherwise if we
just need metric for native balance it's fine to remove
`BalanceCollector` and do it directly in `Signer` future.

I've also added tracing spans for all of the signers