refactor: introduce `SignerTask` and move queue to tx service (#392)
RIght now if signer capacity is reached we are starting to queue
transactions on it. This is not great as we might want to send those
transactions to other signers in cases when their throughput is better
or if our signer is getting stuck/running out of funds.

The solution here is to track queue on the `TransactionService`, and
dispatch queued transactions to the first signer that has capacity.

This is a bit tricky because right now capacity is only available in the
signer future state through `pending.len()`, and the tx service has no
access to it. I could see 2 approaches to solve it
1. Have signers communicate their capacity through events — this could
have been a subject to weird race conditions which would be critical in
this case
2. Track capacity manually on transaction service based on number of
transactions we've sent — This would have required us to also track
additional spawned tasks like loaded transactions and nonce gaps which
would leak signer internals into tx service
3. Refactor signer future to be a nameable type with state we can access
— This seemed like the most reasonable option and also allowed tx
service to have direct access to signer state (e.g to `paused` flag)

This PR introduces `SignerTask` which does the same thing as the current
signer future but is nameable and provides access to `Arc<Signer>`.
Capacity of signer is now available directly through `task.capacity()`
and is not a subject to race conditions as it simply reads the length of
underlying futures set