feat: temporary `wallet_getCallsStatus` (#391)
Like https://github.com/ithacaxyz/relay/pull/331 but a hacked version
that is super fragile, but serves as a stepping stone for Porto
integration. Instead of returning an actual bundle ID, it returns a tx
hash. In order to make `wallet_getCallsStatus` work without any external
changes, the chain ID is obtained in a scuffed way.