feat: add tests for signers getting paused/unpaused (#410)
towards #402 

Adds a new test for tx service that artificially lowers signer balances
via `anvil_setBalances`, and asserts that signers are getting paused and
transactions are getting routed to the active one.

I've added `TransactionServiceConfig` which allows to configure
intervals for nonce/balance checks and max signer throughput. This
config is not exposed in cli and only used in tests — otherwise it's
pretty tricky to test this behavior without the test being flaky or
taking too long