fix: determine bundle status based on userop execution (#454)
Currently `relay_getCallsStatus` determines bundle status based on
whether the tx reverted or not, but it is actually expected for the tx
to not revert (in most cases) even if the userop reverts. The only case
where this is not true is on failed payment.

This PR determines bundle status based on whether the userop reverted or
not.