fix: more correct handling of replacements (#459)
It is possible that even after sending replacement transaction or nonce
gap closing transaction, some of the initially sent transactions will
get confirmed.

This PR fixes this by keeping track of all sent transactions
corresponding to a given userop in database, and checking all of them
when waiting for confirmation.