feat: fetch delegation impl from proxy (#519)
Adds the delegation implementation address to the health endpoint. This
fetches the implementation from the proxy, which both serves as an
actual health check, and also ensures that we never misconfigure it. We
just fetch from the first available chain, since they should be the same
across all chains (for now at least)