fix: better logic for tx replacement (#504)
The current logic for transaction replacement does not account for the
minimum fee bumps expected by transaction pools. Additionally, we're
always setting `maxFeePerGas` to a maximum estimated value making bumps
not possible because even if we only need to bump priority fee, we need
to bump `maxFeePerGas` as well.

This PR introduces smarter logic for tx replacement:
1. When producing quotes, we use latest basefee * 2 as a max value (this
is how it works right now already)
2. When sending initial transaction, we use latest basefee + 20% to
leave space for future fee bumps.
3. Every 2 blocks, we check whether we need to replace a transaction.
Replacement tx is sent if either
a) Latest basefee is higher than the `maxFeePerGas` we have currently
b) Our estimation for priority fee shows that it's 10% higher than the
`maxPriorityFeePerGas` we have currently

The constants are pretty random right now and might require adjustment
based on how we actually perform. I'd expect the replacements to be
pretty rare anyway.

There's another simpler way of addressing this via just setting
`maxFeePerGas == maxPriorityFeePerGas` but this would mean that we are
always paying the maximum possible amount which doesn't leave much space
for further optimizations