feat: add test for transaction service restart (#527)
Adds a simple test that checks that we are able to recover queue and
pending transactions on restart.

For this to work I've changed tx service a bit so that it shuts down
when handle is dropped. Another option here could be to test throught
CLI but this will be less flexible.