feat: restrict number of transactions per EOA (#567)
Right now it's possible to simulate multiple different transactions with
the same nonce and submit them all at once. All of them but one will
fail leaving us with loss due to fees paid for the failing transactions.
Additionally, it is possible to occupy infinite number of queue places
for free.

This PR introduces `TxQueue` which is handling all new transactions, and
checks that they are within limits. For now only one limit is enforced —
max number of queued transactions per EOA (the default is 1)

---------

Co-authored-by: Matthias Seitz <matthias.seitz@outlook.de>