fix: better error handling for sequencer service (#652)
During stress test we've hit a case when sequencer for some reason takes
too long to respond (~30 seconds), and eventually returns `nonce too
low` error because the transaction was succesfully forwarded by our node
and already confirmed.

To mitigate this we want to always return the first succesful response.
With this PR we no longer care about errors at all because any succesful
response means transaction was forwarded to sequencer.