fix: use f64 for `paymentPerGas` (#659)
Right now `paymentPerGas` might get rounded to 0 if token decimals are
low (like e.g in USDC case). This PR mitigates this by using `f64` for
`payment_per_gas` instead and only converting it to a `U256` when
passing into simulator contract.

This should ensure correct `paymentAmount` but might result in a bit
wrongly estimated gas so ideally we add precision for `paymentPerGas` on
simulator level and account for it there.