feat: add an option to query public node for tx confirmation (#685)
ref https://github.com/ithacaxyz/relay/issues/556

Introduces `TxMonitoringHandle` which is supposed to encapsulate all of
our logic for transaction watching.

Right now it is simply querying both our provider and an optionally
configured public node that is assumed to be close to sequencer and get
the block faster.

Eventually it might also get extended with logic for tracking
preconfirmations (flashblocks) and maybe get refactored into an actual
service.