fix: never stop attempting to reconnect (#729)
Right now if node is going down and we have a ws connection to it, we
will fallback to default alloy behavior of retrying 10 times with 3s
interval. This is not great and means that after 30s downtime our
provider will become useless causing all relay requests/tasks to fail.

This PR changes transport configuration so that it never stops retrying
to reconnect, thus making sure that as soon as node is back up we will
reconnect