feat: track `required_funds` on `Quote` (#817)
afaict there's currently no way to determine the amount of funds we're
going to pay to the user after `sendPreparedCalls`, however, this is a
useful context for `InteropService`.

This PR adds `required_funds` field to `Quote`. We should remove it if
there'll be an equivalent field on `Intent` but for now it's not too
invasive