feat: always fetch receipt for confirmed transactions (#816)
Note: This is a breaking database change and makes all existing
confirmed transaction entries invalid, thus those are removed as part of
migration which I believe is OK because we don't query historical
statuses anywhere yet iirc

Knowing that transaction was confirmed but not having a receipt is not
very useful because most of the time we want to know whether the intent
itself was successful. e.g for interop even if transaction confirmed we
might not have been able to charge/pay to the user.

Also, with preconfs getting integrated it might not be straighforward to
determine where to fetch receipt from as it might only be available on a
specific endpoint that got the block faster (or on a flashblocks
endpoint)

This PR changes transaction monitoring logic to always fetch receipt of
a transaction and persist it in database. This is integrated into
`getCallsStatus` endpoint and metrics collection logic as well.