feat: `LiquidityTracker` (#818)
When processing interop bundles, we want to ensure that we will never
start processing a bundle we won't have liquidity for. This protects us
from:
a) UX degradation when we let valid users bundles through but are not
able to handle them resulting in user's funds being locked for some
period of time
b) DoS attacks when a single attacker spams relay with many intents,
which we're then forced to refund

The way I've approached this includes 2 steps:
1. Tracking of pending "locked" liquidity. Those are funds that are
about to be used for some of the pending intents and thus we need some
virtual tracking for those.
2. Tracking of confirmed transactions which allows us to remove funds
from virtual "locked" balances as the changes are already onchain. This
should be aware of confirmation block numbers because otherwise we might
hit weird race conditions, especially when using preconfirmations.

This PR introduces `LiquidityTracker` which implements the above. I've
tried to make it easy to follow but let me know if something doesn't
make sense or you can see a simplification path here