feat: add `interop` flag for `CoinRegistry` entries (#837)
We shouldn't block support for a fee token on us being able to provide
interop liquidity for it. This is already an issue in tests where we
have a lot of erc20s of the same kind for which we can't maintain
liquidity (because we assume single token of every kind right now)

This PR adds an `interop` flag that allows us to specify whether we have
liquidity for a selected token.