fix: always fetch orchestrator implementation from account (#847)
We need to make sure we're always using the orchestrator currently used
by account for the digest domain.

For non-precalls this means we should just use `quote.orchestrator`
which always contains the orchestrator fetched from the account.

For precalls we should fetch the orchestrator from the account right
when generating the digest.