refactor: storage-based liquidity locking (#839)
Based on #826 

To lock same liquidity across different relay instances we need to
somehow synchronize the locking which is happening in memory right now
and thus only takes effect in scope of single relay instance.

Given that we're going to have a single postgres database for all
instances soon (if not already), this PR makes locks storage-based and
performs synchronization via postgresql locks (`select for update`).

This actual change is that logic is moved from `LiquidityTrackerInner`
to `InMemoryStorage` and re-implemented via sql in `PgStorage`.

The most interesting part is here and I'd appreciate if someone with
good understanding of postgresql locks went through it:

https://github.com/ithacaxyz/relay/blob/2e3d72bcbac3ba031ac32fe0855d507d9469ec6d/src/storage/pg.rs#L382-L473