feat: `RebalanceService` (#826)
Introduces `RebalanceService` which is responsible for monitoring
current balances across chains and performing rebalancing if our
balances go below thresholds.

Essentially `RebalanceService` is simply checking the balances and
looking for the largest deviation across chains for a given token. Once
it's found, we rebalance from a chain with highest excess to the chain
with highest shortfall.

---------

Co-authored-by: joshieDo <93316087+joshieDo@users.noreply.github.com>