feat: multichain intent fees (#869)
Implements multichain fees, that is, input intents will now pay fees
associated with escrowing, and output intents will pay fees associated
with the action itself.

This requires us to assess balances on other chains as the balance of
the fee token *minus* the cost of escrowing those tokens. This makes fee
estimation for multichain intents a multistep process.

The way this works is (roughly):

- Estimate the output intent as if it was single chain to get an idea of
the fees needed
- Source the funds, adjusting the balance on other chains to be balance
minus the cost of escrowing those funds
- After we've sourced funds, we now simulate the output intent as a
multichain intent given the funds we sourced, which will make it more
expensive
- Since the output intent is now more expensive, we have to source funds
again. This *might* change the number of input chains, which will
increase the cost of the output intent yet again (because of the merkle
verification), so we do this in a loop until the number of input chains
remains constant.

Closes #835