refactor: remove banxa & onramp (#941)
This is no longer used, is most likely completely outdated for what we'd
need, and for now we are likely just going to have onramp be it's own
service to meet deadlines. We can integrate onramp later on in this code
base once we have a clearer view of what is needed.