feat: cross chain stress test (#921)
Modifies the stress testing tool to take a list of RPC URLs. The fee
token is dispersed on each chain to every account, meaning the account
used to initialize stress testing must hold enough fee tokens to cover
`fee_token_amount * number_of_accounts` on each chain.

The accounts are initialized on the chain corresponding to the first RPC
URL passed, and this is also where they will perform actions. I've
adjusted the call bundle from no-ops to transfers. This means that
initially the intents will be single chain by default, as the accounts
have enough funds to cover 9 transfers, after which they will start
using interop to do the remaining transfers. The expected number of
transfers per account (with the default settings) is about `9 *
number_of_chains`.

The following breaking changes have been made:

- `--relay-url` no longer exists, instead pass one or more `--rpc-url`
that points to a proxyd instance that routes traffic to either the relay
or an Ethereum node.
- `--chain-id` is removed, this is fetched from the first `--rpc-url`
passed. This is treated as the output/destination chain in interop.

Closes #820