fix: simulate using 1 wei (#937)
We simulate to get the gas cost of executing an intent. To get an
accurate estimate, we also transfer some amount of fee tokens out of the
EOA to us. Currently, we estimate how much we think the intent will cost
and pass that to the simulator. This means that the simulation will
actually fail if the user does not have enough balance to cover the fee.

However, as noted in #919, this is not the case for eth, because we (for
historical reasons) override the eth balance of the EOA.

This PR harmonizes the two, which also fixes an edge case in multichain
intents. Now, we simulate using a transfer of 1 wei worth of the fee
token, which should be enough to get an accurate estimate, as the gas
cost of transferring an ERC20 is not proportional to the amount (for
most if not all tokens).

In order to make the simulation always pass, we additionally override
the balance for the fee token by adding 1 wei. This means that if the
user has a balance of e.g. 1 USDC, and wants to transfer the entire
thing, we will simulate using 1 USDC + 1 wei worth of USDC which makes
the simulation pass.

To communicate to the SDK whether the user can actually pay for the
intent or not, a new field is added to the quote: `fee_token_deficit`
which represents how many fee tokens the user is missing to execute the
intent.

Closes https://github.com/ithacaxyz/relay/issues/919