fix: correctly estimate fees for multichain intents (#957)
Rewrites the exit condition for estimating multichain intents from
waiting until # of fund sources not changing to an invariant: `balance +
funds - required_assets - fee >= 0`. It is entirely possible for the #
of fund sources to be stable, but the fee & amounts to shift, so this
was incorrect. The new invariant is based on the current balance on the
destination chain (`balance`), the amount we fund the EOA with using the
funder (`funds`), minus the required assets specified by the SDK
(`required_assets`) and the fee of executing the output intent, if
`required_assets` & the fee token is the same token address.

Additionally, we now source funds *first*, and *then* we simulate the
output intent. After that, we check the invariant described above. If
the invariant does not hold, we loop again, up to 3 times.

---------

Co-authored-by: Arsenii Kulikov <klkvrr@gmail.com>