feat: preconf call status (#1007)
Adds a new status code (201) to indicate that all receipts are
available, but at least one transaction in the bundle is preconfirmed
using e.g. Flashblocks.

A tx is considered preconfirmed if the block number in the receipt is
greater than the current block number of the node we are talking to.