fix: make calldata estimate chopped (#1030)
The calldata changes during simulations for multichain intents due to
the random nature of the signature *and* because fund transfer values
also change. This PR simply makes the calldata cost static as a function
of the calldata size. There are some improvements we can make, which I
highlighted in the comment.