feat: check and get verified user email (#1024)
we haven't been able to locate or find the postgres `superuser` that
would let us run the query to create the view and readonly user in this
PR: https://github.com/ithacaxyz/relay/pull/1015

this PR is the other solution I mentioned earlier for checking and
getting user's verified email. This would be called from a CF Worker.

Frontend calls worker `?address=0x....` -> worker calls this method to
check and retrieve verified email.

`api_key` is a key we generate from the relay side and use from the
worker side. The method is not public so no one knows it exists except
us. The API key is to ensure only the worker can call this method (aka
only the call the method if you have an API key).

---------

Co-authored-by: onbjerg <onbjerg@users.noreply.github.com>