fix: zero balance edge case (#1059)
Right now because of us overriding fee token balance from 0 to 1 during
estimation it is possible that simulation will not account for a cold
storage write increasing the balance.

This PR adds a test and fix for it. The fix is to manually add the
missing amount of gas to the gas estimate