feat!: chain-based config (#1083)
1. Removes the asset registry entirely, so no need for a `registry.toml`
2. Removes separate config for interop and fee tokens.
3. Adds a chain based config where assets & endpoints are configured per
chain
4. Adds support for native tokens other than eth
5. Allows segregating different environments on the same relay (i.e.
testnets and non-testnets can use the same relay)
6. Gets rid of `CoinKind` allowing us to just add tokens without code
changes
7. Simplifies some common patterns in the code base by simply shipping
around the chains config and having helper methods on those (e.g. asset
mapping, providers, ...)
8. It is no longer possible to run the relay without a config. This is
because `--fee-tokens` & `--interop-tokens` & `--endpoints` and so on
are non-sensical with the new config. A possible follow up is to simply
remove most of the CLI and just merge env vars & the config, which would
also simplify a lot of the code related to configs/CLI merging.

**Example config**
Check the config examples
[here](https://github.com/ithacaxyz/relay/blob/9661ed560c607fe02445f53aa28efca7fb08d473/README.md#configuration-examples).

**SDK breaking change**
Note that this also breaks the SDK slightly since I've unified a set of
asset types that were not defined in an ERC, and as such
`wallet_getCapabilities` has a slightly different return type for
tokens. Specifically, `kind` does not exist anymore, instead each token
has a `uid`.

Closes #1058