fix: handle duplicate coin ids in coingecko fetcher (#1103)
If multiple assets defined in the config share the same coin ID, we
would only update the price for one of the assets. This makes the
internal map a map of coin IDs to a list of asset UIDs. We only fetch
the price once, but then update all assets with the new price.

This is especially true on e.g. testnet+mainnet dual setups where you
might have an asset `ethereum` and `test-ethereum` that both map to the
coin ID `ethereum`.