fix: only try multichain if `required_asset` is interop enabled (#1107)
We detected some issues running stress tests with our current staging
configuration, which is a set of chains without interop enabled. The
issue turned out to be the fact that we set `required_funds`, which
leads us into the multichain branches of the relay.

This simply does a check before going into that branch to see if the
asset we're requesting is even interop enabled.

A better fix would be https://github.com/ithacaxyz/relay/issues/948