feat(metrics): add symbol and decimals to balance metric (#1185)
Towards https://github.com/ithacaxyz/relay/issues/1187

Adds labels `symbol` and `decimals` to the metric `balance` to correctly
display balances on the chart for chains with non-ETH native tokens.

On screenshot, that's a balance of 10 POL, not 10 ETH.

<img width="1444" height="469" alt="image"
src="https://github.com/user-attachments/assets/88282c66-7571-4209-b2ac-14f9bcef5f53"
/>

---

The dashboard should be fixed to use the correct `decimals` instead of a
constant `1000000000000000000`, and also display the symbol in
parenthesis. Alternatively, we can repeat the chart per `symbol` to
correctly display all numbers in relation to each other.