feat(rpc): include price info in wallet_getAssets (#1205)
This uses the price oracle to fetch prices for assets that are fee
tokens in `wallet_getAssets`.

I added a new type called `AssetMetadataWithPrice` which includes the
price information, used only for the `wallet_getAssets` response.

Adds an e2e test for this

Fixes https://github.com/ithacaxyz/relay/issues/1179

---------

Co-authored-by: joshieDo <93316087+joshieDo@users.noreply.github.com>