feat: add `enum Intent` and `IntentV04` (#1253)
* closes https://github.com/ithacaxyz/relay/issues/1251
* prep for https://github.com/ithacaxyz/relay/issues/1210

With contracts `0.5.0` we have a new Intent struct. We need to support
both if we want auto upgrades to work.

PR introduces the enum below with getters and builder pattern methods. 
```rust
/// Intent enum with versioned variants.
#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
#[serde(untagged)]
pub enum Intent {
    /// Version 0.4 of the Intent struct.
    V04(IntentV04),
}
```

In `0.5.0`, the orchestrator version of an account will dictate if the
intent is initialized with `Intent::v04` or `Intent::v05`