feat: make `prepare_calls` fee token optional (#1246)
closes https://github.com/ithacaxyz/relay/issues/1230

fear not, most changes are on the e2e tests `env.fee_token ->
Some(env.fee_token)`


If a `fee_token` is not passed, the relay will choose the one with the
highest balance in usd, defaulting to the native token. Once we have
https://github.com/ithacaxyz/relay/issues/1231, we can try and be
smarter about it by also looking at other chains.