feat: support signing with EOA root key (#1373)
Closes https://github.com/ithacaxyz/relay/issues/1126

## Description

This PR makes it possible to use the EOA root key to sign intents
without authorizing that key separately. It's cheaper than adding it as
a custom secp256k1 key, because our account has a fast path for
secp256k1 signatures made by the account EOA
https://github.com/ithacaxyz/account/blob/b51320b423ef0d984a69508343de2ab49997453c/src/IthacaAccount.sol#L500-L503

## No breaking changes

The only user-facing change is making the
`SendPreparedCallsParameters.key` field optional, meaning that it can be
omitted to use the EOA as a signer.

## Tests
Unsure if we need more tests other than eager/lazy account upgrade using
the EOA as a key. Lazy upgrade also tests that we can do a simple
transfer using that key.