feat: separate sequence keys for key related operations (#1357)
ref https://github.com/ithacaxyz/relay/issues/1347

This is an attempt to implement better ordering of key-related precalls.
The way it's implemented is by choosing precall nonce based on which
`keyHash` precall is configuring. For multichain precalls dealing with
`keyHash` nonce sequence would be `c1d0 || keyHash[:22] || nonce`. For
non-multichain precalls it'd be `keyHash[:24] || nonce`

To still support up to 2 keys in `prepareUpgradeAccount`, non-admin key
hash is used as sequence key for the initial precall.

This approach is not perfect, here are some issues with it:
1. Multichain and non-multichain sequences can still conflict. 
2. We still don't have any protection from malicious apps configuring
random nonces on purpose.
3. This introduces the problem of conflicting nonces again. e.g if user
wants to create a precall for revoking the key, and precall for
authorizing it was not yet used, app will need to provide the nonce
directly. This might get solved with #1301