feat: support sourcing multiple assets (#1429)
ref https://github.com/ithacaxyz/relay/issues/1231 testcase 5

Implements support for sourcing multiple assets, primarily focusing on a
usecase of a "main" asset being sourced plus a fee token.

This changes `source_funds` to accept a `Vec` of required assets that
are then sourced across different chains.

Logic for picking a funding chains and fee tokens for them can likely be
improved so marking as draft for now

---------

Co-authored-by: joshieDo <93316087+joshieDo@users.noreply.github.com>