feat: fee payer usage improvements (#1440)
* ref #1441

Adds two fee payer states to the interop bundle state machine for the
following scenarios:

1. if a user wants to do a single chain action and specifies a feePayer:
a. if feePayer can cover in that chain, all good and nothing changes
b. if feePayer only has funds on another chain, it creates a interop
bundle with: 1 intent/transaction on chain A for the fee payer and one
transaction/intent on the destination chain for the user. The user
intent will have its payment set to 0 and will only be executed once the
feePayer transaction has succeeded. Note that there's no settlement
and/or sourcing transactions in this scenario.

2. if a user wants to do an interop intent and specifies a feePayer, the
interop bundle will have:
a. 1 transaction/intent on a supported chain for the fee payer, which
covers every user transaction.
b. Once the above completes, we execute the interop bundle as normal.
All payment fields should be set to 0.

breaking changes for SDK:
* adds `feePayerDigest` to response capabilities. `feePayerDigest`
should now be used. It will be equal to the users digest ONLY if it's a
single chain intent in which the fee payer has funds in the same chain.
Otherwise, it will be the digest of a `feePayerQuote`.
* it introduces a `feePayerQuote` field on `Quotes`, that will be filled
when:
   * user has requested an interop intent with a `feePayer`.
* user has requested a single chain intent with a `feePayer`, but the
`feePayer` only has funds in another chain.
* the newly snap files in the PR diff showcase the new fields.